added fix + test
